make explicit future handling in observer instead of using latch
